fix(canon): keep the template instance form off legacy Vue 2 dialects
A legacy Vue 2 default export is a plain options object, not a
constructor, so the public-instance conditional degraded to `{}` and
turned every filter and runtime-resolved template name into a TS2339
the pinned parity surfaces do not contain (vue2-elm, vue-element-admin
across all three dialect configs). The instance form now additionally
requires `!legacy_vue2`, and the gate lives in generate_undefined_refs
beside the emission it guards.

Also restores the source-length budget: the generation entry wrappers
move to generator/entry.rs and the instance-form tests to
tests/options_api_instance.rs, so generator.rs, closures.rs and
tests.rs are all back at or under their merge-base line counts.
